#6 Refactorizacion principal de Ficheros y  LlamadasSiistema
Borramos Encirptado y lo incorporamos a Ficheros, se modifican por rebote LectorMercado y Bot
